+++ app/routes/posts/$id.remove.tsx
@@ -4,6 +4,7 @@ import { deletePostById, getPostById } from "~/models/post.server";
 import { Link, useLoaderData } from "@remix-run/react";
 import type { Post } from "@prisma/client";
 import type { ActionFunction } from "@remix-run/node";
+import { getLocalAuthorizedUser } from "utils/user.server";
 
 export const loader: ActionFunction = async ({ params }) => {
   const id = params.id;
@@ -13,11 +14,20 @@ export const loader: ActionFunction = async ({ params }) => {
   return json<Post>({ ...post });
 };
 
-export const action: ActionFunction = async ({ params }) => {
+export const action: ActionFunction = async ({ request, params }) => {
   const id = params.id;
   invariant(id, "Invalid id");
-  await deletePostById(id);
-  return redirect("../");
+  const user = await getLocalAuthorizedUser(request);
+  if (!user) {
+    return redirect("/login");
+  }
+  if (
+    user.role === "admin" ||
+    user.role === "contributor" ||
+    user.role === "mod"
+  )
+    await deletePostById(id);
+  return redirect("/lists");
 };
 export default function RemoveListById() {
   const { title } = useLoaderData<Post>();
